use int64 as dtype of shape input

+++ onnx/backend/test/case/node/reshape.py
@@ -16,11 +16,11 @@ class Reshape(Base):
     def export():
         original_shape = [2, 3, 4]
         test_cases = {
-            'reordered_dims': np.array([4, 2, 3], dtype=np.int32),
-            'reduced_dims': np.array([3, 8], dtype=np.int32),
-            'extended_dims': np.array([3, 2, 2, 2], dtype=np.int32),
-            'one_dim': np.array([24], dtype=np.int32),
-            'negative_dim': np.array([6, -1, 2], dtype=np.int32),
+            'reordered_dims': np.array([4, 2, 3], dtype=np.int64),
+            'reduced_dims': np.array([3, 8], dtype=np.int64),
+            'extended_dims': np.array([3, 2, 2, 2], dtype=np.int64),
+            'one_dim': np.array([24], dtype=np.int64),
+            'negative_dim': np.array([6, -1, 2], dtype=np.int64),
         }
         data = np.random.random_sample(original_shape).astype(np.float32)
